+++ fdm-calculator/src/bln3/input.ts
